Renamed findKeywordInPageTitle to findKeyphraseInSEOTitle.
